Standardize timestamp format on Creator Analysis tab
**Change:** Updated timestamp format from 'short' month to 'numeric' month to match User Analysis tabs.

**Before:** "Last updated: Jan 15, 2025 3:45 PM"
**After:** "Last updated: 1/15/2025, 3:45 PM"

Ensures consistent timestamp display across all 4 analysis tabs.

+++ creator_analysis_tool_supabase.js
@@ -308,7 +308,7 @@ class CreatorAnalysisToolSupabase extends CreatorAnalysisTool {
             // Save results to localStorage
             const now = new Date();
             const timestamp = now.toLocaleString('en-US', {
-                month: 'short',
+                month: 'numeric',
                 day: 'numeric',
                 year: 'numeric',
                 hour: 'numeric',
@@ -446,7 +446,7 @@ class CreatorAnalysisToolSupabase extends CreatorAnalysisTool {
             // Save results to localStorage
             const now = new Date();
             const timestamp = now.toLocaleString('en-US', {
-                month: 'short',
+                month: 'numeric',
                 day: 'numeric',
                 year: 'numeric',
                 hour: 'numeric',
